Refactor auth routes into separate area
