+++ auth-service/routes/auth/auth-forgot-password.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { createApp } from "../app.ts";
+import { createApp } from "../../app.ts";
 import { testRateLimiting } from "./test-helpers.ts";
 import { EmailClient } from "@saflib/email";
