+++ auth-service/routes/auth/auth-logout.ts
