+++ auth-service/routes/auth/auth-resend-verification.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { createApp } from "../app.ts";
+import { createApp } from "../../app.ts";
 import passport from "passport";
 import { testRateLimiting } from "./test-helpers.ts";
 import { EmailClient } from "@saflib/email";
