+++ auth-service/routes/auth/auth-reset-password.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import request from "supertest";
 import express from "express";
-import { createApp } from "../app.ts";
+import { createApp } from "../../app.ts";
 import { testRateLimiting } from "./test-helpers.ts";
 // Mock argon2
 import passport from "passport";
