+++ auth-service/routes/auth/auth-reset-password.ts
