+++ auth-service/routes/auth/auth-verify.ts
